updating documentation and moving S3ObjectStorageClassFilter.java to s3-api and some minor changes
