+++ hadoop-tools/hadoop-aws/src/site/markdown/tools/hadoop-aws/index.md
@@ -926,8 +926,18 @@ The switch to turn S3A auditing on or off.
   <description>
     Should auditing of S3A requests be enabled?
   </description>
-</property>
+</property
+```
+## <a name="glacier"></a> Glacier Object Support
 
+[Amazon S3 Glacier (S3 Glacier)](https://docs.aws.amazon.com/amazonglacier/latest/dev/introduction.html) is a secure and durable service for low-cost data archiving and
+long-term backup.
+With S3 Glacier, you can store your data cost effectively for months, years, or even decades.
+This support introduces a new config, which decides the objects returned from listStatus. 
+Note : This is not available on all AWS S3 store types, or on third party stores.
+
+The config is as follows: 
+```xml
 <!--
 The switch to control how S3A handles glacier storage classes.
 -->
@@ -943,6 +953,8 @@ The switch to control how S3A handles glacier storage classes.
 </description>
 </property>
 ```
+
+
 ## <a name="retry_and_recovery"></a>Retry and Recovery
 
 The S3A client makes a best-effort attempt at recovering from network failures;
